Added Authorize Middleware

+++ routes.php
@@ -1,20 +1,18 @@
 <?php
-
-
 $router->get('/', 'HomeController@index');
 $router->get('/listings', 'ListingController@index');
-$router->get('/listings/create', 'ListingController@create');
+$router->get('/listings/create', 'ListingController@create', ['auth']);
+$router->get('/listings/edit/{id}', 'ListingController@edit', ['auth']);
+$router->get('/listings/search', 'ListingController@search');
 $router->get('/listings/{id}', 'ListingController@show');
-$router->post('/listings', 'ListingController@store');
-$router->delete('/listings/{id}', 'ListingController@destroy');
-$router->get('/listings/edit/{id}', 'ListingController@edit');
-$router->put('/listings/{id}', 'ListingController@update');
 
-$router->get('/auth/register', 'UserController@create');
-$router->get('/auth/login', 'UserController@login');
+$router->post('/listings', 'ListingController@store', ['auth']);
+$router->put('/listings/{id}', 'ListingController@update', ['auth']);
+$router->delete('/listings/{id}', 'ListingController@destroy', ['auth']);
 
-$router->post('/auth/register', 'UserController@store');
+$router->get('/auth/register', 'UserController@create', ['guest']);
+$router->get('/auth/login', 'UserController@login', ['guest']);
 
-$router->post('/auth/logout', 'UserController@logout');
-
-$router->post('/auth/login', 'UserController@authenticate');
+$router->post('/auth/register', 'UserController@store', ['guest']);
+$router->post('/auth/logout', 'UserController@logout', ['auth']);
+$router->post('/auth/login', 'UserController@authenticate', ['guest']);
